+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/dcBotModule.kt
@@ -5,6 +5,7 @@ import io.github.sophon.adminModule
 import io.github.sophon.core.coreModule
 import io.github.sophon.discord.data.InMemoryGlossaryDB
 import io.github.sophon.discord.featureRegistry.featureRegistryModule
+import io.github.sophon.discord.usecase.CreateErrorEmbedUseCase
 import io.github.sophon.discord.usecase.RouteCommandToFeatureUseCase
 import io.github.sophon.dreamcancel.dreamCancelModule
 import io.github.sophon.glossaryinfil.data.GlossaryDB
@@ -51,7 +52,9 @@ fun dcBotModule(kord: Kord) = module {
     single { kord }
 
     singleOf(::DiscordBotImpl).bind<DiscordBot>()
+
     singleOf(::RouteCommandToFeatureUseCase)
+    singleOf(::CreateErrorEmbedUseCase)
 
     singleOf(::InMemoryGlossaryDB).bind<GlossaryDB>()
 }