+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/domain/DiscordRegisteredFeature.kt
@@ -1,6 +1,5 @@
 package io.github.sophon.discord.domain
 
-import dev.kord.rest.builder.message.EmbedBuilder
 import io.github.sophon.core.domain.Result
 import io.github.sophon.core.feature.FeatureInfo
 import io.github.sophon.core.feature.Game
@@ -24,42 +23,4 @@ internal interface DiscordRegisteredFeature {
         query: String,
         origin: Source,
     ): Result<BotOutput, BotError>
-}
-
-data class SupportedCommand(
-    val command: Command,
-    val description: String,
-    val arguments: List<Argument> = listOf(),
-) {
-    data class Argument(
-        val name: String,
-        val description: String,
-        val isRequired: Boolean = true,
-    )
-}
-
-data class BotOutput(
-    val embedBuilder: (EmbedBuilder.() -> Unit)? = null,
-    val plainText: String? = null,
-    val errorEmbedBuilder: (EmbedBuilder.() -> Unit)? = null,
-    val images: Images? = null,
-    val feedback: Feedback? = null,
-    val reply: Reply? = null,
-) {
-    data class Images(
-        val title: String,
-        val titleUrl: String,
-        val urls: List<String>,
-    )
-
-    data class Feedback(
-        val embedBuilder: (EmbedBuilder.() -> Unit),
-        val origin: Source,
-        val feedbackChannelList: List<String>,
-    )
-
-    data class Reply(
-        val embedBuilder: (EmbedBuilder.() -> Unit),
-        val target: Source,
-    )
 }