+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/featureRegistry/featureRegistryModule.kt
@@ -3,7 +3,6 @@ package io.github.sophon.discord.featureRegistry
 import io.github.sophon.core.feature.Config
 import io.github.sophon.discord.config.ConfigLoader
 import io.github.sophon.discord.domain.DiscordRegisteredFeature
-import io.github.sophon.discord.domain.FeatureRegistry
 import io.github.sophon.discord.domain.Scheduler
 import io.github.sophon.discord.featureRegistry.admin.AdminDiscordFeature
 import io.github.sophon.discord.featureRegistry.admin.usecase.BanUseCase