+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/usecase/CreateCharacterAliasesEmbedUseCase.kt
@@ -4,23 +4,29 @@ import dev.kord.rest.builder.message.EmbedBuilder
 import io.github.sophon.core.domain.Result
 import io.github.sophon.core.domain.map
 import io.github.sophon.core.domain.mapError
+import io.github.sophon.core.feature.FeatureInfo
 import io.github.sophon.core.wiki.domain.WikiClient
 import io.github.sophon.core.wiki.domain.model.Character
 import io.github.sophon.discord.BotError
 import io.github.sophon.discord.domain.toDomainError
+import io.github.sophon.discord.util.featureFooter
 import io.github.sophon.discord.util.mandatoryField
 
 internal class CreateCharacterAliasesEmbedUseCase {
-    suspend fun invoke(wiki: WikiClient): Result<EmbedBuilder.() -> Unit, BotError> {
+    suspend fun invoke(
+        wiki: WikiClient,
+        featureInfo: FeatureInfo,
+    ): Result<EmbedBuilder.() -> Unit, BotError> {
         return wiki.fetchCharacterList()
             .mapError { it.toDomainError() }
             .map { characterList ->
-                createAliasesEmbed(characterList)
+                createAliasesEmbed(characterList, featureInfo)
             }
     }
 
     private fun createAliasesEmbed(
         characterList: List<Character>,
+        featureInfo: FeatureInfo,
     ): EmbedBuilder.() -> Unit = {
         val string = buildString {
             characterList
@@ -35,5 +41,7 @@ internal class CreateCharacterAliasesEmbedUseCase {
             name = "🥸 CHARACTER ALIASES",
             value = string,
         )
+
+        featureFooter(featureInfo)
     }
 }