+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # BOT CHANGELOG
 
+## [v3.0.0] - 2025-01-02
+- added footers to all wiki Discord embeds
+- invalid queries include syntax embeds
+- `alias` variants for KOF and COTW
+
 ## [v2.1.1] - 2025-12-31
 - Wavu v1.0.8 - added TekkenDocs mention